% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Geometric Mean}
\usage{
geomean(x)
}
\arguments{
\item{x}{Vector of positive numbers.}
}
\description{
This function allows you to calculate the geometric mean of a vector of positive numbers.
}
\examples{
geomean(c(2.2,5.3,7,1,0.2))
}
\keyword{geometric}
\keyword{mean}
