% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_install.R
\name{library.install}
\alias{library.install}
\title{Adaptive Package Installation and Loading}
\usage{
library.install(vec.of.packages)
}
\arguments{
\item{vec.of.packages}{Character vector of package names.}
}
\description{
This function allows you to load packages already installed and install those missing.
}
\examples{
library.install(c("dplyr","knitr"))
library.install("knitr")
}
\keyword{install}
\keyword{library}
