% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelplot_anyKind.R
\name{panelplot.anyKind}
\alias{panelplot.anyKind}
\title{Creation of a Matrix of Many Plots for Exploratory Analyses}
\usage{
panelplot.anyKind(tdat, ynames = colnames(tdat), xnames = colnames(tdat),
  tnrows = 2, tncols = 2, maxperpage = tnrows * tncols,
  new.page.x.var = FALSE, new.page.y.var = FALSE, plotmain = NULL,
  uniqueplots = FALSE, skipSameXY = FALSE, save.plots = FALSE,
  type.plots = "pdf", latex = FALSE, full.latex = FALSE,
  basename = NULL, lookup.vec = NULL, showr = TRUE, use.pch = 1,
  use.cex = 1, wait.between = FALSE, ...)
}
\arguments{
\item{tdat}{dataframe with variables for y-axes,}

\item{ynames}{names of variables to go on y-axis}

\item{xnames}{names of variables to go on x-axis}

\item{tnrows}{number of rows of plots}

\item{tncols}{number of columns of plots}

\item{plotmain}{title to go at top of each page of plots}

\item{uniqueplots}{TRUE means only make unique plots}

\item{skipSameXY}{TRUE means skip plots that have the same X and Y variable; only applies when uniqueplots=FALSE}

\item{save.plots}{TRUE means to save plots in a file}

\item{type.plots}{"pdf"  to create pdf files (only if save.plots=T), "eps" to create eps files}

\item{latex}{TRUE means create LaTeX code to show plots, and have includegraphics line}

\item{full.latex}{TRUE if want latex output to be a stand-alone file}

\item{basename}{base name of saved plots}

\item{lookup.vec}{vector of longer names, where column names should match variable names in tdat}

\item{showr}{TRUE to include correlation and n in plot heading}

\item{use.pch}{pch for plotting (in scatter.plot)}

\item{use.cex}{cex for plotting (in scatter.plot)}

\item{wait.between}{TRUE if want to wait between screens of plots}

\item{...}{Other graphical parameters sent to plot}
}
\description{
Creates one to a large number of exploratory plots, where
the user has control over how the subplots (panels) are arranged
on the plots, without needing to keep track of how many plots
are created.  The function can be used to display the plots on
the screen, pausing between each plot, or to save the plots
to either pdf of eps files.  If the latter, the user also has
the option to create a simple LaTeX fragment which can be included
in LaTeX code to show all the plots created.  The scatter plots
by default show the correlation and sample size on the plot titles
(the user can change this).  By giving the relevant argument for
lookup.vec, this function will call on lookup to display the longer
names for the x- and y-axes.
There are 2 primary situations in which one might want to call
panelplot: (1) when we want to plot one or more outcome variables
(ynames) versus one or more covariates (xnames); and (2) when we
want to make all unique plots.  The latter is similar to pairs(),
but it doesn't include duplicate plots, the user has control over
how many plots to place on a page, and the plots can contain the
correlations and number of observations on the plot title.
}
\examples{
test.dat<-data.frame(x=rnorm(10),y=rbinom(10,2,.5))
panelplot.anyKind(test.dat)

test.dat<-data.frame(x=1:10+rnorm(10),y=as.factor(c(rep(1,5),rep(0,5))))
panelplot.anyKind(test.dat)
}
\keyword{layout}
\keyword{scatter}
\keyword{scatterplot}
