% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_date.R
\name{paste.date}
\alias{paste.date}
\title{File Names Using Current Dates}
\usage{
paste.date(...)
}
\arguments{
\item{...}{A character string intended to be the file name.}
}
\description{
Function returns a character string consisting of the user-supplied character string and the current date in %Y-%m$d format.
}
\examples{
paste.date("Todays Analysis.Rmd")
}
\keyword{date}
