% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_box_mosaic_plot.R
\name{scatter.box.mosaic.plot}
\alias{scatter.box.mosaic.plot}
\title{Quick Adaptive Plotting Function for Exploratory Analyses}
\usage{
scatter.box.mosaic.plot(tdat, yname, xname, long.yname = NULL,
  long.xname = NULL, showr = TRUE, use.pch = 1, use.cex = 1, jitx = 0,
  jity = 0, addtomain = NULL, cex.xylab = 1.5, cex.title = 1.2, ...)
}
\arguments{
\item{yname}{name of variable from tdat to plot on y-axis}

\item{xname}{name of variable from tdat to plot on x-axis}

\item{long.yname}{character string to be put on y-axis}

\item{long.xname}{character string to be put on x-axis}

\item{showr}{TRUE means include correlation and n on plot heading}

\item{use.pch}{pch symbol to be used in plot}

\item{use.cex}{cex (character expansion) for plotting symbols}

\item{jitx}{factor component of jitter(x,factor=jitx)}

\item{jity}{factor component of jitter(y,factor=jity)}

\item{addtomain}{additional text for main title (if showr=FALSE, this is the entire title)}

\item{cex.xylab}{cex for x and y axes}

\item{cex.title}{cex for title (and correlation, if shown)}

\item{...}{Other graphical parameters sent to plot}
}
\description{
Function to make many scatterplots, boxplots and mosaic plots as necessary for exploratory analyses. Includes the Spearman correlation coefficient and sample size in the plot title. Correlation coefficients for plots with any categorical variables are calculated after converting the factors to numeric. The titles are printed in red if the correlation is significant. Significance is tested by the method appropriate to the variable: scatterplots use Spearman's test, boxplots use Kruskal-Wallis and mosaics use Chi-Square test.
}
\examples{
test.dat<-data.frame(x=rnorm(10),y=rbinom(10,2,.5))
scatter.box.mosaic.plot(test.dat,yname="y",xname="x")

test.dat<-data.frame(x=1:10+rnorm(10),y=as.factor(c(rep(1,5),rep(0,5))))
scatter.box.mosaic.plot(test.dat,yname="y",xname="x")
}
\keyword{scatter}
\keyword{scatterplot}
