% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.plot.R
\name{scatter.plot}
\alias{scatter.plot}
\title{Quick Scatterplot Function for Exploratory Analyses}
\usage{
scatter.plot(tdat, yname, xname, long.yname = NULL, long.xname = NULL,
  showr = TRUE, use.pch = 1, use.cex = 1, jitx = 0, jity = 0,
  addtomain = NULL, cex.xylab = 1.5, cex.title = 1.2, ...)
}
\arguments{
\item{yname}{name of variable from tdat to plot on y-axis}

\item{xname}{name of variable from tdat to plot on x-axis}

\item{long.yname}{character string to be put on y-axis}

\item{long.xname}{character string to be put on x-axis}

\item{showr}{TRUE means include correlation and n on plot heading}

\item{use.pch}{pch symbol to be used in plot}

\item{use.cex}{cex (character expansion) for plotting symbols}

\item{jitx}{factor component of jitter(x,factor=jitx)}

\item{jity}{factor component of jitter(y,factor=jity)}

\item{addtomain}{additional text for main title (if showr=FALSE, this is the entire title)}

\item{cex.xylab}{cex for x and y axes}

\item{cex.title}{cex for title (and correlation, if shown)}

\item{...}{Other graphical parameters sent to plot}
}
\description{
Function to make many scatterplots for exploratory analyses. Includes Pearson correlation and sample size in the plot title.
}
\examples{
test.dat<-data.frame(x=rnorm(10),y=rbinom(10,2,.5))
scatter.plot(test.dat,yname="y",xname="x")
}
\keyword{scatter}
\keyword{scatterplot}
