% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensSpec.R
\name{SensSpec}
\alias{SensSpec}
\title{Sensitivity and Specificity}
\usage{
SensSpec(x, grouping, thresh = seq(0, 1, 0.05), plots = T, table = F,
  legend.loc = c(0.7 * max(thresh), 0.85))
}
\arguments{
\item{x}{Vector of predicted probabilities for classification. For example, the predicted probabilities from a logistic model.}

\item{grouping}{The vector of true class labels for the data.}

\item{thresh}{Vector of desired thresholds for classification. The table will have as many rows as thresholds.}

\item{plots}{Logical, plot the sensitivity, specificity and error rate curves versus the threshold vector on one plot.}

\item{table}{Logical, returns the table of sensitivities, specificities and error rates for all thresholds.}

\item{legend.loc}{Vector of two values for the x and y coordinate of the legend, respectively. Applicable only when plot is TRUE.}
}
\description{
Function to calculate a table of sensitivities, specficities and error rates for classification procedures using predicted probabilities and true classification labels.
}
\examples{
test.dat<-data.frame(x=rnorm(10),y=rbinom(10,2,.5))
SensSpec(test.dat$x,test.dat$y)
SensSpec(test.dat$x,test.dat$y,plot=FALSE,tab=TRUE)
}
\keyword{classification}
\keyword{error}
\keyword{sensitivity}
\keyword{specificity}
